% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGlmnet.R
\name{plotBetas}
\alias{plotBetas}
\title{Plot the coefficients from glmnet}
\usage{
plotBetas(glmnet, labelLambda = 0, ylab = "Coefficient",
  transformFunc = list(function(x) x, function(x) x), ...)
}
\arguments{
\item{glmnet}{a glmet object from \code{\link{glmet}}}

\item{labelLambda}{label all variables that are not 0 at this the closest lambda <= labelLambda}

\item{ylab}{label for y axis}

\item{transformFunc}{a function to transform the betas by (e.g. \code{exp})}

\item{...}{additional arguments for \code{\link{plot}}}
}
\description{
Take a glmnet object and plot out the path of the coefficient for each variable as model complexity increases.
}
\examples{
#example from cv.glmet
library(glmnet)
set.seed(1010)
n=1000;p=100
nzc=trunc(p/10)
x=matrix(rnorm(n*p),n,p)
beta=rnorm(nzc)
fx= x[,seq(nzc)] \%*\% beta
eps=rnorm(n)*5
y=drop(fx+eps)
px=exp(fx)
px=px/(1+px)
ly=rbinom(n=length(px),prob=px,size=1)
set.seed(1011)
cvob1=cv.glmnet(x,y)
plotBetas(cvob1$glmnet.fit,cvob1$lambda.1se)
}

