% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGlmnet.R
\name{plotGlmnet}
\alias{plotGlmnet}
\title{Plot model performance from cv.glmnet}
\usage{
plotGlmnet(fit, markBest1SE = FALSE, ...)
}
\arguments{
\item{fit}{a cv.glmet object from \code{\link[glmnet]{cv.glmnet}}}

\item{markBest1SE}{add a vertical line through the least complex model within 1 standard error of the best model}

\item{...}{extra arguments to plot}
}
\description{
Take a cv.glmnet object and plot out the change in performance as model complexity increases
}
\examples{
#example from cv.glmet
set.seed(1010)
n=1000;p=100
nzc=trunc(p/10)
x=matrix(rnorm(n*p),n,p)
beta=rnorm(nzc)
fx= x[,seq(nzc)] \%*\% beta
eps=rnorm(n)*5
y=drop(fx+eps)
px=exp(fx)
px=px/(1+px)
ly=rbinom(n=length(px),prob=px,size=1)
set.seed(1011)
cvob1=glmnet::cv.glmnet(x,y)
plotGlmnet(cvob1)
}
